Update ofxOculusDK2.cpp
I was trying ofxOculusDK2 on with oF 0.9 RC2 (on OS X 10.10.5 with Xcode 7.1), and found the `hmd->DisplayId` field is not valid (therefore accessing `displays[whichDisplay]` on the next line crashes).  It looks like this field may have been scheduled for deletion.

I simply implemented what your comment suggests (take 2nd one if there are multiple displays, otherwise use the first).  This works with Oculus runtime 0.5.0.1.